Trait rustless::backend::Handler

pub trait Handler {
    fn call(&self, &mut Request) -> Result<Response, Box<Error + 'static>>;

    fn catch(&self, &mut Request, err: Box<Error + 'static>) -> (Response, Result<(), Box<Error + 'static>>) { ... }
}

Handlers are responsible for handling requests by creating Responses from Requests.

By default, bare functions and variants of Chain implement Handler.

Handlers are allowed to return errors, and if they do, their catch method is called and the error is propagated to AfterMiddleware.

Required Methods

fn call(&self, &mut Request) -> Result<Response, Box<Error + 'static>>

Produce a Response from a Request, with the possibility of error.

If this returns an Err, catch is called with the error.

Provided Methods

fn catch(&self, &mut Request, err: Box<Error + 'static>) -> (Response, Result<(), Box<Error + 'static>>)

If Handler's call method produces an Err, then this method is called to produce a Response and possibly handle the error.

If the passed-in error is not handled, it should be returned as the second item in the returned tuple. If it is handled, then Ok(()) can be returned instead to indicate that all is good with the Response and the error has been dealt with.