Module rustless::batteries::swagger

Structs

Contact
Contact information for the exposed API.

Info
The object provides metadata about the API. The metadata can be used by the clients if needed, and can be presented in the Swagger-UI for convenience.

License
License information for the exposed API.

Spec
This is the root document object for the API specification.

SwaggerSpecKey

Enums

Scheme
The transfer protocol for the operation. Values MUST be from the list: "http", "https", "ws", "wss". The value overrides the Swagger Object schemes definition.

Functions

build_spec
Build the basic Swagger 2.0 object

create_api
Create an API to handle doc requests

enable