Trait rustless::framework::api_handler::ApiHandler

pub trait ApiHandler: Typeable {
    fn api_call<'a, 'b>(&'a self, &str, &mut Json, &'b mut Request + 'b, &mut CallInfo<'a>) -> HandleResult<Response>;
}

Required Methods

fn api_call<'a, 'b>(&'a self, &str, &mut Json, &'b mut Request + 'b, &mut CallInfo<'a>) -> HandleResult<Response>

Methods

impl ApiHandler

fn is<E: ApiHandler>(&self) -> bool

Is this Error object of type E?

fn downcast<E: ApiHandler>(&self) -> Option<&E>

If this error is E, downcast this error to E, by reference.

fn downcast_mut<T: ApiHandler>(&mut self) -> Option<&mut T>

Returns some mutable reference to the boxed value if it is of type T, or None if it isn't.

unsafe fn downcast_mut_unchecked<T: ApiHandler>(&mut self) -> &mut T

Returns a mutable reference to the boxed value, blindly assuming it to be of type T. If you are not absolutely certain of T, you must not call this.