Trait rustless::framework::ApiHandler

pub trait ApiHandler: Any {
    fn api_call<'a, 'b>(&'a self, &str, &mut Json, &'b mut Request + 'b, &mut CallInfo<'a>) -> HandleResult<Response>;
}

Required Methods

fn api_call<'a, 'b>(&'a self, &str, &mut Json, &'b mut Request + 'b, &mut CallInfo<'a>) -> HandleResult<Response>

Methods

impl ApiHandler

fn is<T: ApiHandler>(&self) -> bool

Returns true if the boxed type is the same as T

fn downcast_ref<T: ApiHandler>(&self) -> Option<&T>

Returns some reference to the boxed value if it is of type T, or None if it isn't.

unsafe fn downcast_ref_unchecked<T: ApiHandler>(&self) -> &T

Returns a reference to the boxed value, blindly assuming it to be of type T. If you are not absolutely certain of T, you must not call this.

fn downcast_mut<T: ApiHandler>(&mut self) -> Option<&mut T>

Returns some mutable reference to the boxed value if it is of type T, or None if it isn't.

unsafe fn downcast_mut_unchecked<T: ApiHandler>(&mut self) -> &mut T

Returns a mutable reference to the boxed value, blindly assuming it to be of type T. If you are not absolutely certain of T, you must not call this.

impl ApiHandler

fn downcast<T: ApiHandler>(self: Box<Self>) -> Result<Box<T>, Box<Self>>

Returns the boxed value if it is of type T, or Err(Self) if it isn't.

unsafe fn downcast_unchecked<T: ApiHandler>(self: Box<Self>) -> Box<T>

Returns the boxed value, blindly assuming it to be of type T. If you are not absolutely certain of T, you must not call this.