Struct rustless::server::header::AccessControlRequestHeaders

pub struct AccessControlRequestHeaders(pub Vec<String>);

The Access-Control-Request-Headers request header, part of CORS.

The Access-Control-Request-Headers header indicates which headers will be used in the actual request as part of the preflight request.

Spec: www.w3.org/TR/cors/#access-control-request-headers-request-header