Trait rustless::server::header::HeaderFormatExperimental

pub trait HeaderFormat: Send + Sync + Clone + Any {
    fn fmt_header(&self, fmt: &mut Formatter) -> Result<(), Error>;
}

A trait for any object that will represent a header field and value.

This trait represents the formatting of a Header for output to a TcpStream.

Required Methods

fn fmt_header(&self, fmt: &mut Formatter) -> Result<(), Error>

Format a header to be output into a TcpStream.

This method is not allowed to introduce an Err not produced by the passed-in Formatter.