Module rustless::server::net

A collection of traits abstracting over Listeners and Streams.

Structs

HttpConnector
A connector that will produce HttpStreams.

NetworkConnections
An iterator wrapper over a NetworkAcceptor.

Enums

Fresh
The write-status indicating headers have not been written.

HttpListener
A NetworkListener for HttpStreams.

HttpStream
A wrapper around a TcpStream.

Streaming
The write-status indicating headers have been written.

Traits

NetworkConnector
A connector creates a NetworkStream.

NetworkListener
An abstraction to listen for connections on a certain port.

NetworkStream
An abstraction over streams that a Server can utilize.

Type Definitions

ContextVerifier