Trait rustless::server::net::NetworkAcceptor

pub trait NetworkAcceptor: Clone where <Self as NetworkAcceptor>::Stream: Send, <Self as NetworkAcceptor>::Stream: Sized, <Self as NetworkAcceptor>::Stream: NetworkStream, <Self as NetworkAcceptor>::Stream: Clone {
    type Stream;
    fn accept(&mut self) -> Result<<Self as NetworkAcceptor>::Stream, IoError>;
    fn socket_name(&self) -> Result<SocketAddr, IoError>;
    fn close(&mut self) -> Result<(), IoError>;

    fn incoming(&mut self) -> NetworkConnections<Self> { ... }
}

An abstraction to receive NetworkStreams.

Required Methods

type Stream

fn accept(&mut self) -> Result<<Self as NetworkAcceptor>::Stream, IoError>

Returns an iterator of streams.

fn socket_name(&self) -> Result<SocketAddr, IoError>

Get the address this Listener ended up listening on.

fn close(&mut self) -> Result<(), IoError>

Closes the Acceptor, so no more incoming connections will be handled.

Provided Methods

fn incoming(&mut self) -> NetworkConnections<Self>

Returns an iterator over incoming connections.