Trait rustless::server::net::NetworkListener

pub trait NetworkListener: Clone {
    type Stream: Send + NetworkStream + Clone;
    fn accept(&mut self) -> Result<Self::Stream, Error>;
    fn local_addr(&mut self) -> Result<SocketAddr, Error>;

    fn incoming(&mut self) -> NetworkConnections<Self> { ... }
}

An abstraction to listen for connections on a certain port.

Associated Types

type Stream: Send + NetworkStream + Clone

The stream produced for each connection.

Required Methods

fn accept(&mut self) -> Result<Self::Stream, Error>

Returns an iterator of streams.

fn local_addr(&mut self) -> Result<SocketAddr, Error>

Get the address this Listener ended up listening on.

Provided Methods

fn incoming(&mut self) -> NetworkConnections<Self>

Returns an iterator over incoming connections.