Module rustless::server::server

HTTP Server

Example

use hyper::server::{Server, Request, Response};
use hyper::status::StatusCode;
use hyper::uri::RequestUri;

let server = Server::http(|req: Request, mut res: Response| {
    *res.status_mut() = match (req.method, req.uri) {
        (hyper::Get, RequestUri::AbsolutePath(ref path)) if path == "/"  => {
            StatusCode::Ok
        },
        (hyper::Get, _) => StatusCode::NotFound,
        _ => StatusCode::MethodNotAllowed
    };
}).listen("0.0.0.0:8080").unwrap();

Modules

request
Server Requests

response
Server Responses

Structs

Listening
A listening server, which can later be closed.

Request
A request bundles several parts of an incoming NetworkStream, given to a Handler.

Response
The outgoing half for a Tcp connection, created by a Server and given to a Handler.

Server
A server can listen on a TCP socket.

Enums

Fresh
The write-status indicating headers have not been written.

Streaming
The write-status indicating headers have been written.

Traits

Handler
A handler that can handle incoming requests for a server.