Struct rustless::server::server::response::Response

pub struct Response<'a, W = Fresh> {
    pub version: HttpVersion,
    // some fields omitted
}

The outgoing half for a Tcp connection, created by a Server and given to a Handler.

Fields

version

The HTTP version of this response.

Methods

impl<'a, W> Response<'a, W>

fn status(&self) -> StatusCode

The status of this response.

fn headers(&self) -> &Headers

The headers of this response.

fn construct(version: HttpVersion, body: HttpWriter<&'a mut Writer + 'a>, status: StatusCode, headers: Headers) -> Response<'a, Fresh>

Construct a Response from its constituent parts.

fn deconstruct(self) -> (HttpVersion, HttpWriter<&'a mut Writer + 'a>, StatusCode, Headers)

Deconstruct this Response into its constituent parts.

impl<'a> Response<'a, Fresh>

fn new(stream: &'a mut Writer + 'a) -> Response<'a, Fresh>

Creates a new Response that can be used to write to a network stream.

fn start(self) -> Result<Response<'a, Streaming>, IoError>

Consume this Response, writing the Headers and Status and creating a Response

fn status_mut(&mut self) -> &mut StatusCode

Get a mutable reference to the status.

fn headers_mut(&mut self) -> &mut Headers

Get a mutable reference to the Headers.

impl<'a> Response<'a, Streaming>

fn end(self) -> Result<(), IoError>

Flushes all writing of a response to the client.