Struct rustless::server::server::Server

pub struct Server<L = HttpListener> {
    // some fields omitted
}

A server can listen on a TCP socket.

Once listening, it will create a Request/Response pair for each incoming connection, and hand them to the provided handler.

Methods

impl Server<HttpListener>

fn http(ip: IpAddr, port: u16) -> Server<HttpListener>

Creates a new server that will handle HttpStreams.

fn https(ip: IpAddr, port: u16, cert: Path, key: Path) -> Server<HttpListener>

Creates a new server that will handler HttpStreamss using a TLS connection.

impl<L, A, S> Server<L> where L: Send, L: NetworkListener, <L as NetworkListener>::Acceptor == A, A: Send, A: NetworkAcceptor, <A as NetworkAcceptor>::Stream == S, S: Send, S: NetworkStream, S: Clone

fn with_listener(ip: IpAddr, port: u16, listener: L) -> Server<L>

Creates a new server that will handle HttpStreams.

fn listen_threads<H>(self, handler: H, threads: usize) -> Result<Listening<<L as NetworkListener>::Acceptor>, HttpError> where H: Handler

Binds to a socket, and starts handling connections using a task pool.

fn listen<H>(self, handler: H) -> Result<Listening<<L as NetworkListener>::Acceptor>, HttpError> where H: Handler

Binds to a socket and starts handling connections.