Module rustless::server_backend::header::commonExperimental

Common Headers A Collection of Header implementations for common HTTP Headers.

Mime

Several header fields use MIME values for their contents. Keeping with the strongly-typed theme, the mime crate is used, such as ContentType(pub Mime).

Modules

accept
Exposes the Accept header.

connection
Exposes the Connection header.

content_length
Exposes the ContentLength header.

content_type
Exposes the ContentType header.

date
Exposes the Date header.

host
Exposes the Host header.

location
Exposes the Location header.

server
Exposes the Server header.

transfer_encoding
Exposes the TransferEncoding header.

user_agent
Exposes the UserAgent header.

Structs

Accept
The Accept header.

ContentLength
The Content-Length header.

ContentType
The Content-Type header.

Date
The Date header field.

Host
The Host header.

Location
The Location header.

Server
The Server header field.

TransferEncoding
The Transfer-Encoding header.

UserAgent
The User-Agent header field.

Enums

Connection
The Connection header.