Module rustless::server::server

HTTP Server

Modules

request
Server Requests

response
Server Responses

Structs

Fresh
The write-status indicating headers have not been written.

Listening
A listening server, which can later be closed.

Request
A request bundles several parts of an incoming NetworkStream, given to a Handler.

Response
The outgoing half for a Tcp connection, created by a Server and given to a Handler.

Server
A server can listen on a TCP socket.

Streaming
The write-status indicating headers have been written.

Traits

Handler
A handler that can handle incoming requests for a server.