Struct rustless::server::server::Server

pub struct Server<'a, H, L = HttpListener> where H: Handler {
    // some fields omitted
}

A server can listen on a TCP socket.

Once listening, it will create a Request/Response pair for each incoming connection, and hand them to the provided handler.

Methods

impl<'a, H, L> Server<'a, H, L> where L: NetworkListener, H: Handler

fn new(handler: H) -> Server<'a, H, L>

Creates a new server with the provided handler.

impl<'a, H> Server<'a, H, HttpListener> where H: 'static + Handler

fn http(handler: H) -> Server<'a, H, HttpListener>

Creates a new server that will handle HttpStreams.

fn https(handler: H, cert: &'a Path, key: &'a Path) -> Server<'a, H, HttpListener>

Creates a new server that will handler HttpStreamss using a TLS connection.

fn https_with_context(handler: H, ssl_context: SslContext) -> Server<'a, H, HttpListener>

Creates a new server that will handler HttpStreamss using a TLS connection defined by an SslContext.

impl<'a, H> Server<'a, H, HttpListener> where H: 'static + Handler

fn listen_threads<T>(self, addr: T, threads: usize) -> Result<Listening, HttpError> where T: ToSocketAddrs

Binds to a socket, and starts handling connections using a task pool.

fn listen<T>(self, addr: T) -> Result<Listening, HttpError> where T: ToSocketAddrs

Binds to a socket and starts handling connections.

impl<'a, H, L, S> Server<'a, H, L> where H: 'static + Handler, S: Send + NetworkStream + Clone, L: Send + 'static + NetworkListener<Stream=S>

fn with_listener(self, listener: L, threads: usize) -> Result<Listening, HttpError>

Creates a new server that will handle HttpStreams.