Struct rustless::server::server::Server

pub struct Server<'a, H, L = HttpListener> where H: Handler {
    // some fields omitted
}

A server can listen on a TCP socket.

Once listening, it will create a Request/Response pair for each incoming connection, and hand them to the provided handler.

Methods

impl<'a, H, L> Server<'a, H, L> where H: Handler, L: NetworkListener

fn new(handler: H) -> Server<'a, H, L>

impl<'a, H> Server<'a, H, HttpListener> where H: 'static, H: Handler

fn http(handler: H) -> Server<'a, H, HttpListener>

Creates a new server that will handle HttpStreams.

fn https(handler: H, cert: &'a Path, key: &'a Path) -> Server<'a, H, HttpListener>

Creates a new server that will handler HttpStreamss using a TLS connection.

impl<'a, H> Server<'a, H, HttpListener> where H: 'static, H: Handler

fn listen_threads(self, ip: IpAddr, port: u16, threads: usize) -> Result<Listening, HttpError>

Binds to a socket, and starts handling connections using a task pool.

fn listen(self, ip: IpAddr, port: u16) -> Result<Listening, HttpError>

Binds to a socket and starts handling connections.

impl<'a, H, L, S> Server<'a, H, L> where H: 'static, H: Handler, L: Send, L: 'static, L: NetworkListener, <L as NetworkListener>::Stream == S, S: Send, S: NetworkStream, S: Clone

fn with_listener(self, listener: L, threads: usize) -> Result<Listening, HttpError>

Creates a new server that will handle HttpStreams.